requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
appdirs==1.4.4
async-timeout==4.0.3
attrs==23.2.0
black==24.2.0
certifi==2024.2.2
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Deprecated==1.2.14
diff_cover==8.0.3
dune_client==1.4.0
exceptiongroup==1.2.0
frozenlist==1.4.1
idna==3.6
iniconfig==2.0.0
Jinja2==3.1.3
MarkupSafe==2.1.5
multidict==6.0.5
mypy-extensions==1.0.0
ndjson==0.3.1
numpy==1.26.4
packaging==23.2
pandas==2.2.0
pathspec==0.12.1
platformdirs==4.2.0
pluggy==1.4.0
Pygments==2.17.2
pytest==8.0.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
six==1.16.0
tblib==3.0.0
toml==0.10.2
tomli==2.0.1
tqdm==4.66.2
types-Deprecated==1.2.9.20240106
types-python-dateutil==2.8.19.20240106
types-PyYAML==6.0.12.12
types-requests==2.31.0.20240218
types-setuptools==69.1.0.20240217
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.2.1
wrapt==1.16.0
yarl==1.9.4
